% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonotoneBayes-package.R
\docType{package}
\name{MonotoneBayes-package}
\alias{MonotoneBayes-package}
\alias{MonotoneBayes}
\title{The 'MonotoneBayes' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
