% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotoneBayes.R
\name{monotoneBayes}
\alias{monotoneBayes}
\title{Title}
\usage{
monotoneBayes(
  X,
  Y,
  L = 10,
  tau0 = 1e-04,
  c_sq = 10^2,
  fix = F,
  c_alpha = 0.01,
  c_beta = 0.01 * 4,
  prior = "Regularized HS"
)
}
\arguments{
\item{X}{input x vector}

\item{Y}{input binary response variable}

\item{L}{input the number of user-defined interval}

\item{tau0}{input hyperparameter}

\item{c_alpha}{input hyperparameter for c}

\item{c_beta}{input hyperparameter for c}

\item{prior}{choose the shrinkage prior to use, can be "Regularized HS", "Orginal HS", "Laplacian" and "Gaussian"}
}
\value{
a list
$model: the model result
$plots: the fitted curve with original dataset
}
\description{
Title
}
\examples{
X <- runif(10, 0, 1)
Y <- rbinom(10, 1, 0.2)
monotoneBayes(X, Y)

}
