% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCurve.R
\name{predictCurve}
\alias{predictCurve}
\title{March6:
The predictCurve returns a simplified posterior mean/median curves which
linear connected the posterior mean/median or node points.}
\usage{
predictCurve(model, L, method = "mean", grid.eq = T, node.t, newdata)
}
\arguments{
\item{model}{Fitted Model}

\item{L}{the pre-defined L for model}

\item{method}{posteiror mean or median used}

\item{grid.eq}{equal grid, if F need to provide node.t below}

\item{node.t}{default equal nodes on 0-1}

\item{newdata}{newdata}
}
\value{
a vector of fitted prob at given x
}
\description{
Unfinished: the case grid.eq == F hasn't been finish.
}
