% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModel.R
\name{sumStats}
\alias{sumStats}
\title{sumStats}
\usage{
sumStats(model, AllPara = F, ndigit = 3, orgHS = F)
}
\arguments{
\item{model}{rstan object, including the result from monotoneBayes()}

\item{AllPara}{default F, if T then return all parameter in the model including c,tau,and lambda}

\item{ndigit}{default 3, the digits of results}

\item{orgHS}{default F, only T when the model is fitted via original HS method}
}
\value{
The summary statistics of fitted model
}
\description{
sumStats
}
