% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotoneBayes.R
\name{monotoneBayes}
\alias{monotoneBayes}
\title{Title}
\usage{
monotoneBayes(
  X,
  Y,
  L = 10,
  tau0 = 0.01,
  nodes = seq(0, 1, length.out = 10 + 1),
  Eq.Space = T,
  c_sq = 10^2,
  fix = F,
  c_alpha = 1,
  c_beta = 1 * 4,
  prior = "Regularized HS",
  ...
)
}
\arguments{
\item{X}{input x vector}

\item{Y}{input binary response variable}

\item{L}{input the number of user-defined interval}

\item{tau0}{input hyperparameter}

\item{nodes}{default 0,1/L,2/L,...L/L: user-defined nodes point}

\item{Eq.Space}{default T, should be F when user defined interval not equal space}

\item{c_sq}{fixed C_sq value for regularized HS with fixed C.}

\item{fix}{default F}

\item{c_alpha}{input hyperparameter for c}

\item{c_beta}{input hyperparameter for c}

\item{prior}{choose the shrinkage prior to use, can be "Regularized HS", "Orginal HS", "Laplacian" and "Gaussian"}

\item{...}{chain parameter passing to stan, inclcuding iter, warmup, control ,...}
}
\value{
a list
$model: the model result
$plots: the fitted curve with original dataset
}
\description{
Title
}
\examples{
X <- runif(10, 0, 1)
Y <- rbinom(10, 1, 0.2)
monotoneBayes(X, Y)

}
