% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModel.R
\name{plotModel}
\alias{plotModel}
\title{plotModel}
\usage{
plotModel(model, L, method = "mean", grid.eq = T, node.t)
}
\arguments{
\item{model}{rstan object, including the result from monotoneBayes()}

\item{L}{user-defined interval}

\item{method}{fitted curve use either posteiror "mean" or "median"}

\item{grid.eq}{default T for equal spacing}

\item{node.t}{if grid.eq == T, equally spaced nodes at 0-1, or when grid.eq = F, this should be user-defined node point}
}
\value{
a list
$FittedProb: a data frame list the estimated value at specific points of \link{0,1}
$plots: the fitted model curve
}
\description{
plotModel
}
