% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{predictModel}
\alias{predictModel}
\title{Title
Mar 6: this is corrected version of predict function: return posterior mean/median and 95\% crediable interval}
\usage{
predictModel(model, L, method = "mean", grid.eq = T, node.t, newdata)
}
\arguments{
\item{model}{Fitted Model}

\item{L}{the pre-defined L for model}

\item{method}{posteiror mean or median used}

\item{grid.eq}{equal grid, if F need to provide node.t below}

\item{node.t}{default equal nodes on 0-1}

\item{newdata}{newdata}
}
\value{
a vector of fitted prob at given x
}
\description{
Title
Mar 6: this is corrected version of predict function: return posterior mean/median and 95\% crediable interval
}
